Infrastructure records are now created in MongoDB before server starts
Instead of using the YML file we have an object representation of the data that we can use throughout the app.

I split out the jobs of loading the YML file and creating the records. Not only does it ensure single responsibility but it allows a rake task to be written that doesn't have to know about how the data is loaded. We can ask it to "CreateInfrastructureRecords" and trust it to figure it out. Keeping the `FindDalmatianConfiguration` class separate also helps us stub out different fixtures in the test suite.

The Infrastructure fields are set to `Mongoid::Attributes::Dynamic` for now. We know the dalmatian.yml file is prone to change so a level of flexiblity is likely going to be helpful to start this app off. We can come back and tighten the expected interface later.