Users can view individual infrastructures for more detail
Currently this work surfaces all of the infrastructure config from the YML file. The main goal is to reproduce the environment variable management feature so it's imagined that this page will iterate later but that displaying the content we have is easy enough to do quickly to form a meaningful show page.