Run script/test through Docker
We want to use the same set of scripts to run the test suite regardless of environment.

We must use Docker on CI and in production environments https://github.com/dxw/tech-team-rfcs/blob/main/rfc-013-use-docker-to-deploy-and-run-applications-in-containers.md

The scripts/ do not work straight away when envoked through docker. Packages like yarn, rbenv and shellcheck are not available but the script/test is expecting them to be there. To make this work I've grouped the 2 setups into 2. Either the CI ENV is present in which case we defer to the Dockerfile to manage the dependencies OR we assume we are running locally without docker and must interact with the native machine, where packages like rbenv may be present.

We could add a ENV["DOCKER"] flag in as well but as Docker must be used on CI it may be ok given the comment.