CI can test the application with Docker Compose
Docker must be used for testing the application on CI https://github.com/dxw/tech-team-rfcs/blob/main/rfc-013-use-docker-to-deploy-and-run-applications-in-containers.md

GitHub actions was configured to test with Postgres as per the
rails-template. This needs updating to use Mongo.

After some trouble fixing Docker related problems outside of this PR I
created a docker-compose.yml file to use locally to replicate the issue.
It captures the instructions and dependencies in a reusable
docker-compose command that may be valuable again in future.

Docker compose continues to ensure script/test is the single command for
testing the application inline with this dxw RFC[1].

I chose to use environment variables for DATABASE_URL for test in the
same way as development to make it consistent and configurable rather
than hardcoded, keeping close to the principles of 12-factor.

[1] https://github.com/dxw/tech-team-rfcs/blob/main/rfc-023-use-scripts-to-rule-them-all.md