Set AWS Dalmatian profile if present
This needs to be specificed within the context of each server process to allow the AWS API to authenticate with the correct user. If this is not done, the default profile is used which may not be Dalmatian.

For those users who have Dalmatian as their default profile nothing new will run and it should work out of the box.

If this is not done authentication errors pop up immediately.