App requires an environment variable for AWS_ROLE
To keep this configurable and to avoid hardcoding dalmatian specific language, a new environment will be responsibile for providing the role name.

CI fails without a change to the .dockerignore to include the test env
file. This allows Dotenv to find a key and value it can use for running
the test suite.