FindEnvironmentVariable service can fetch all variables for an infrastructure
Each infrastructure may include information about services and environments, it may not. Where they have been included we know that we should be attempting to retrieve environment variables for a service running on each environment.

There may not be a service deployed to both environments specified yet. I think an empty table that shows no environment variables is better than hiding it completely since it allows the user to see the request was made but it came up empty.

The 'role_session_name' name is odd. In the previous prototype it was hardcoded to this despite an apparent effort to set it to something more specific [1]. Removing it is not an option since it is a required argument but the value I provide seems to have no bearing on the results. I'm leaving it in for now incase it does become relevant.

[1] https://github.com/dxw/dalmatian-gui/blob/170c189296d8f048a38bb042b601d4770a772c7e/app/controllers/dalmatian_parameter_store_controller.rb#L5