Refactor AWS API spec helpers so that client stubbing is decoupled
Before we add in support for the stubbing of a new API method, refactor the client stubbing into their own methods for easier reuse.

The aws_sts_client can remain private however stub_aws_ssm_client will need to be public for the logic in find_environment_variables_spec. It requires that 4 requests are stubbed (for each service and environment), to so this we need to ensure the client is the same double. Exposing stub_aws_ssm_client as a public method allows this to happen.