Refactor where the page header for environment variables lives
I wanted to add a link to add a new variable below the title rather than above it. At this point I noticed that the _table partial was doing more than rendering a table, it was also responsible for the header. Instead of making the situation worse by moving the links into the _table partial, I'm moving the header out.