Users can add a new environment variable to an infrastructure
It became apparent during this work that the add and update API methods are the same. What changes is whether the variable name matches an existing environment variable or not. I've added a note onto the new form to explain this. This commit includes only the scenario for adding a new variable, an update spec will follow.

Given we really want to make sure we are only applying the environment variable to the correct infrastructure, service and environment. We check the params for service_name and environment_name before using them, making sure they are included in the infrastructures known entities.