Throw away checks to ensure variables are only deleted for the intended service
As there is no formal data association between Infrastructure and EnvironmentVariable the state is passed around with query params.

Due to this choice I had added some checks to try to assert if those query params had been changed. For example, if the create or delete action was called for staging but somehow changed to be applied to production, guard against it.

In order to make this approach work for the #delete action as well as the #create action I set off down a rabbit hole of trying to use a before_action. Given how the parameters are provided in different ways this quickly led me to question the need for this logic at all. I have instead added the service name and environment name that is in the form request, into the confirmation box. This will hopefully provide confidence and solve the problem of catching any unexpected code changes which breaks the way we send this state around.