EnvironmentVariables can be asked for their full name
When we send the request to AWS we prefix the name with the service and environment. Presumably to ensure they are unique and can be identified when looking at parameter store.

Using a method centralises the way this value is calcuated so it can be used consistently between the create and destroy journeys.