CI can run feature tests that require javascript
Install Firefox (by the name of iceweasel on Ubuntu) so that we can use
selenium_headless as the Selenium driver. This will allow the test suite
to run and check javascript when specifed by js: true.

For non JS tests the default Rack::Test is still used as the driver.
This is the default choice of Capybara but I've made it explicit so the
distinction is clearer to the reader.

Including extra packages like Firefox into the core Dockerfile is not
ideal. It means it is always present in running containers and as well
as taking more time to build, could present a larger surface area for
unexpected and unknown exploits. Given the way the script/test and
script/bootstrap files are written to use brew rather than using Docker
to run the tests, on CI this is problematic. Installing Brew takes a
long time and cannot be cached unlike a Docker layer. Given Docker is
only used for CI it felt the best compromise to add this as a dependency
for the docker image that places the firefox binary into position for
Selenium to use.

Install Firefox and geckodriver in the Brewfile so the tests can still be run locally should the user not have firefox installed.