CI uses Firefox whilst dev uses Chrome for browser tests
Firefox was working on CI but I missed the fact that it started breaking outside of docker.

The error included Geckodriver in the middle of several chrome references. Odd since I thought the `selenium_headless` driver was Firefox by default [1].

```
  Failure/Error: Unable to find WebDriverError@chrome://marionette/content/error.js to read failed line

     Selenium::WebDriver::Error::InvalidArgumentError:
       Expected "handle" to be a string, got [object Undefined] undefined
     # WebDriverError@chrome://marionette/content/error.js:175:5
     # InvalidArgumentError@chrome://marionette/content/error.js:304:5
     # assert.that/<@chrome://marionette/content/assert.js:479:13
     # assert.string@chrome://marionette/content/assert.js:383:53
     # GeckoDriver.prototype.switchToWindow@chrome://marionette/content/driver.js:1645:10
     # despatch@chrome://marionette/content/server.js:305:40
     # execute@chrome://marionette/content/server.js:275:16
     # onPacket/<@chrome://marionette/content/server.js:248:20
     # onPacket@chrome://marionette/content/server.js:249:9
     # _onJSONObjectReady/<@chrome://marionette/content/transport.js:501:20
```

[1] https://github.com/teamcapybara/capybara#selenium

chromium-chromedriver is not available in the debian packages so I cannot figure out how to get chrome running on CI. Given I've been stuck on this for a time I'm going to leave it with Chrome for dev and Firefox for CI. The driver should be interchangeable from Capybara's point of view. Docker tests can always be run locally too if a problem with parity comes up.