Environment variables are now shown on their own index page
Move from the infrastructure show action as the page was getting over crowded, especially for infrastructures with several services and environments.

Use standard bootstrap tabs.

Putting environment variables in the second tab rather than the first. Even though the Dalmatian config as it is isn't that helpful at the moment I imagine this first tab will become a more generic info page with an assortment of helpful links. There is more chance of the Dalmatian config moving down the pecking order to the last tab when this happens, but we'll see what the users need when we get there.