Test coverage is sent to Coveralls by a GitHub action
Change from Coveralls gem to simplecov because it automatically tries to send during the rspec command. This would fail on CI since it has no access to a token.

Instead of creating a coveralls.yml and trying to safely add a secret there (given the tests are running in a container too), we change approach to a github action.

The github action [1] allows us to run after the tests have been run in an explict step. We can then provide the `GITHUB_TOKEN` to it via Github's `secrets` interface.

Generating the lcov file was a pain. Coveralls and Simplecov create a json file and don't appear to have any way to configure them to lcov - something the github action requires.

We use a simplecovlsov gem to format the output into the required format.

The test suite runs in the docker container and stores the output in a new file. Given this is in a container we need to share this file back to the file system/context of where github actions is running. We do this with a little volume mapping. We can then provide the file path to the github action and it knows which file to report.

[1] https://github.com/coverallsapp/github-action/