Users can see infrastructure variables for each environment
Following the steer of the GUI prototype which used the path of "/dalmatian-variables/infrastructures/<infrastructure>/<environment>/". Using this path successfully retrieves existing values so that is why it's used.

Another difference between these and the other environment variables is that these are not concerned with the service. This diverges what's required at both service and view level so whilst there is some duplication it will allow these differently managed variables to be extended in different ways.