Fix partial so it receives local variables
Tightening the interface for a partial makes it easier to reuse. By reaching out and looking for `@instance_variable` for one instance but being passed the local environment variables for the others makes things available in 2 ways.

To reuse a partial like this would then require the reader to stumble onto this problem when it broke, even after seemingly honouring the contract of the local variables.