Refactor fetching of environment variables
FindEnvironmentVariables and FindInfrastructureVariables services both contained the same code required to create a request to AWS for parameters (as they call them).

Separate the responsibility of working out what requests need to be made, which paths we need to ask AWS for and how to organise the results, from the making of the AWS request. This allows both services to reuse a single definition, reducing duplication and increasing maintainability.