Users can create or update infrastructure variables
Not to be confused with "environment variables". These variables sit at an infrastructure level rather than per service. An example of this would be a `GITHUB_TOKEN` which would need to be available to AWS Codepipeline in other to run CI tasks and report statuses to github. These variables aren't required for a service to run, but are required to facilitate their auxilary tasks.