Users can view and hide the values of secret environment variables
The table cell obfuscated value starts off by being padded by whitespace. To avoid double clicking the cell for the toggle to work we strip the whitespace off of the value to make a reliable value to compare against.

Use coffeescript as we already have a ga.coffee and coffeescript is found referenced from Rails documentation [1], implying a convention which I'm happy to follow.

The user experience for this could be improved. It isn't clear that clicking the asterisks does anything or indeed clicking them again restores it to being hidden. Adding a button into the table cell is also not brilliant. Feedback from the show and tell was that it was workable for now and there are only a few users who will be using it. The next step would probably be to add an explifit show/hide all button at the top of the page and leave this feature in.

[1] https://guides.rubyonrails.org/working_with_javascript_in_rails.html#page-change-events