Interacting with infrastructure variables uses different AWS client
Environment variables are managed from the AWS account associated with the infrastructure. Infrastructure variables are managed from a central and single AWS account.

Though we use our local AWS credentials through `Aws::STS::Client.new` to authenticate connections, we can also use it to grab the account_id of the core/main AWS account. This is the value we need to use when constructing the `role_arn` for any request that is related to infrastructure variables.

AWS services (eg PutAwsParameter) now expect a `client` to be created by the service that chooses to use them. This allows these classes to be reused in the new case of having to use 2 different clients to connect with. This method will also allow other clients to be specified in future. It will also allow us to change back to a single client if we choose to move where infrastructure variables exist within AWS.