Users can upload a file of environment variables to add/update
Uploaded files become temp files that Ruby only persists for the duration of the request. In order to have a confirmation step where the file must be retrieveable in order to pass to the creation step, we copy the tempfile into our local tmp directory.

The tmp folder is excluded from git tracking as other files are such as capybara screenshots. We may still wish to tidy these files up to limit the room for accidents with those secrets being copied into another direction and accidentally committed, etc.