Handle missing values on multiple variable upload
Use a new service called CreateMultipleEnvironmentVariables to manage the task of uploading more than 1 variable at a time. This gives us a opportunity to reuse CreateEnvironmentVariable whilst being able to wrap it in some error handling.

The error tested with was a missing value however this approach should handle any type of `Aws::SSM::Errors::ValidationException` and bubble up the message to the user.