Validate uploaded ENV files
Try to use regex to spot if each line is as expected. This isn't strictly required as AWS will error and it will bubble up however it can be quite a lot of information. To help the user with the most common errors validate against the rules we specify to the user and let them know if it passed or failed.

We could do more to say which line failed and why in future if this is needed.