Rename upload button to make room for another button
"Upload multiple variables" is perhaps too verbose given the buttons are always within the context of an "environment variable" tab and their properties.

This change makes room for a counter "Download" button that  if it followed this precedent would need to be something like "Download all variables". A little too wordy.