Clean local ENV downloads on new session
Whenever the server is started the tmp/downloads folder is removed and added again as an empty directory.

This is done primarily to mitigate the risk of any user storing up all sorts of sensitive environment variables on their machine, should their machine be compromised or that directory accidentally committed.

A secondary issue is that there could be many hundreds of these files if the tool is used a lot and we should have some automated way to clean these up. On the basis that the these downloaded files are served to the client there is already a copy being stored separately by the user, removing these should have no impact.