Display a flash error when the local AWS token has expired
I was unable to find a way to test this given the way the AWS gem initiliases itself.

When we write `Aws::STS:Client.new` with no arguments the gem will go off and load the credentials from your file system [1]. What it will also do is reinitialise itself after loading in these aguments. To successfully stub a fake exception here I would either have to pass in the credentials directly and lose parity with how it's really done or I'd have to dig deeper into the gem in terms of stubbing. This latter option took me down an unsuccessful rabbit hole.

[1] https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/STS/Client.html#initialize-instance_method