ENV changes prompt the user into started a new deploy
ENV changes do not automatically get provisions to running instances of the service.

Without this change the changed or new variables wouldn't be available to application code until a deployment had been done.

Now that we have code pipelines and the ability to execute new runs we can at least point the user to that so that they can choose whether or not to redeploy.

I have not taken the step of automatically starting a new build for 2 reasons:

1. It's not always clear what pipeline a user would need to run. Everything to do with pipelines is tied to the name which should be constructed of infrastructure, service and environment names but this is not guaranteed.
2. A user (especially if adding one at a time) may not wish to do a deployment at the same time or at least want to make all their changes first to avoid many deploys.

Later we could change this to make it automatic.