Core Dalmatian AWS account ID configurable
Instead of asking AWS to return the ID of the central account when setting up the clients, we expect this value to be injected before run time.

This should make working with this value easier in the back end as we look to create more connection solely to AWS with this account_id.