Try to set DALMATIAN_AWS_ACCOUNT_ID on boot
The environment variable DALMATIAN_AWS_ACCOUNT_ID corresponds to the main AWS account. This account should be what users are set up with locally to use this service and others such as Dalmatian Tools.

Given these is a convention for setting all users up in this way we can try to be smart about setting this variable where it's absent.

It does require an extra API request but it is fast so the trade off for convienence is something we think is acceptable for.

If a user does want to configure differently the user is able to overwrite this by providing one.