Set DALMATIAN_CI_PIPELINE on boot if not present
We have since learnt that this string is hardcoded into Dalmatian as the core dalmatian ci that deploys terraform across all infrastructures.

Given this name will be the same for all users of dalmatian it seems like we can get some value of convienence by setting it on server start.

In case this does change or needs to be overwritten we assign it to ENV to keep _this_ app configurable.