Add local copy of AWS MFA script
Dalmatian Tools introduced a breaking change that removed this method and changed approach [1]. Instead of writing to the `~/.aws/*` files began writing to `$HOME/.config/dalmatian/*`, a custom file. These credentials are unavailable to this application which is looked for by the AWS CLI and AWS Gem which looks in the _default_ credentials location: `~/.aws/*`.

Some time was spent trying to copy credentials from the new file Dalmatian creates back into the default location so things _just work_. It is hard to rewrite the ~/.aws/credentials text file every 12 hours by adding a new mfa block but not changing anything else, this is fragile, and why I believe Dalmatian Tools changed approach.

We cannot load the vales from the new dalmatian credentials into ENV because the AWS gem doesn't have an MFA environment variable that I could find. It only has an access_key and secret_token. This may be a missing part of the gem but even if we could, the AWS CLI is also used in the start up script which would need provisioning with env correctly too. There may be a way to do this but copying the old working script so that a developer can have the new version of both repos is the fastest thing we can do.

[1] https://github.com/dxw/dalmatian-tools/pull/36