Repo now works again by requiring Dalmatian Tools as a hard dependency
The start script had become out of date and didn't work to authenticate properly for AWS. Making the app completely unusable.

Dalmatian Tools [1] is now more stable and feature rich. It is the single place that is being actively worked on so it makes sense to try to reuse it and having a single definition for local authenticattion with AWS.

In terms of changes to the start script we use 2 bits of magic from Dalmatian Tools:

1. Whenever we use the `dalmatian` binary it calls the bin/aws/assume-infrastructure-role script [2]. This looks at your local AWS configuration found in `.aws` and/or `.config/dalmatian/` to connect and assume the right AWS role. If you've run `dalmatian login` and gone through that process prior to this, this should work behind the scenes
2. We use `dalmatian util exec` to wrap our commands in a way that includes the necessary environment variables needed for AWS Authentication from within a Rails context: `ENV["AWS_ACCESS_KEY"]` etc.

One note on the use of `dalmatian exec`. Given we pass the output of this to `jq` we have to make sure we strip any comments from it with `sed '/==>/d'`. Every comment should start "==>" so while fragile it should be good enough to hook onto this convention for now.

[1] https://github.com/dxw/dalmatian-tools
[2] https://github.com/dxw/dalmatian-tools/blob/main/bin/dalmatian#L263