No longer automatically fix Ruby linting problems on test run
It's not clear what changes have and haven't been caused by Rubocop during development. Instead of correcting things aggresively with automatic fixes, we fail if it breaks and prompt the user to understand what about the linting was wrong. The Ruby style guides should be standard and not project specific so once we are practiced in the right style then we'll write the code in the expected way.

We can always run `bundle exec standardrb --fix` manually and squash those changes into the appropriate commit.